pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "diffprompt"
version = "0.1.0"
description = "git diff for your prompt's behavior"
authors = [{ name = "Rudra Dudhat", email = "contact.rdudhat@gmail.com" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["llm", "prompt", "testing", "evaluation", "diff"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Testing",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "sentence-transformers>=2.2.0",
    "hdbscan>=0.8.29",
    "umap-learn>=0.5.0",
    "httpx>=0.24.0",
    "pydantic>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
diffprompt = "diffprompt.cli:app"

[project.urls]
Homepage = "https://github.com/RudraDudhat2509/diffprompt"
Repository = "https://github.com/RudraDudhat2509/diffprompt"

[tool.hatch.build.targets.wheel]
packages = ["diffprompt"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
